// src/meta/primevue/types.ts
export enum ComponentType {
  AutoComplete = "AutoComplete",
  CascadeSelect = "CascadeSelect",
  Checkbox = "Checkbox",
  ColorPicker = "ColorPicker",
  DatePicker = "DatePicker",
  Editor = "Editor",
  FloatLabel = "FloatLabel",
  IconField = "IconField",
  InputGroup = "InputGroup",
  InputMask = "InputMask",
  InputNumber = "InputNumber",
  InputOtp = "InputOtp",
  InputText = "InputText",
  KeyFilter = "KeyFilter",
  Knob = "Knob",
  Listbox = "Listbox",
  MultiSelect = "MultiSelect",
  Password = "Password",
  RadioButton = "RadioButton",
  Rating = "Rating",
  Select = "Select",
  SelectButton = "SelectButton",
  Slider = "Slider",
  Textarea = "Textarea",
  ToggleButton = "ToggleButton",
  ToggleSwitch = "ToggleSwitch",
  TreeSelect = "TreeSelect",
  Button = "Button",
  SpeedDial = "SpeedDial",
  SplitButton = "SplitButton",
  DataTable = "DataTable",
  DataView = "DataView",
  OrderList = "OrderList",
  OrgChart = "OrgChart",
  Paginator = "Paginator",
  PickList = "PickList",
  Timeline = "Timeline",
  Tree = "Tree",
  TreeTable = "TreeTable",
  VirtualScroller = "VirtualScroller",
  Accordion = "Accordion",
  Card = "Card",
  Deferred = "Deferred",
  Divider = "Divider",
  Fieldset = "Fieldset",
  Panel = "Panel",
  ScrollPanel = "ScrollPanel",
  Splitter = "Splitter",
  Stepper = "Stepper",
  Tabs = "Tabs",
  Toolbar = "Toolbar",
  ConfirmDialog = "ConfirmDialog",
  ConfirmPopup = "ConfirmPopup",
  Dialog = "Dialog",
  Drawer = "Drawer",
  DynamicDialog = "DynamicDialog",
  Popover = "Popover",
  Tooltip = "Tooltip",
  Upload = "Upload",
  Breadcrumb = "Breadcrumb",
  ContextMenu = "ContextMenu",
  Dock = "Dock",
  Menu = "Menu",
  Menubar = "Menubar",
  MegaMenu = "MegaMenu",
  PanelMenu = "PanelMenu",
  TieredMenu = "TieredMenu",
  Chart = "Chart",
  Message = "Message",
  Toast = "Toast",
  Carousel = "Carousel",
  Galleria = "Galleria",
  Image = "Image",
  ImageCompare = "ImageCompare",
  AnimateOnScroll = "AnimateOnScroll",
  Avatar = "Avatar",
  Badge = "Badge",
  BlockUI = "BlockUI",
  Chip = "Chip",
  FocusTrap = "FocusTrap",
  Inplace = "Inplace",
  Fluid = "Fluid",
  MeterGroup = "MeterGroup",
  ScrollTop = "ScrollTop",
  Skeleton = "Skeleton",
  ProgressBar = "ProgressBar",
  ProgressSpinner = "ProgressSpinner",
  Ripple = "Ripple",
  StyleClass = "StyleClass",
  Tag = "Tag",
  Terminal = "Terminal",
}